JDBC compatibility fixes
Fixes #1. Changes to Java 8 specifications removed daylight time standards from JDBC in releases. When a daylight code was active on the connection this would crash JDCServer at startup and prevent database access if the connection had already been initialized. This fix adds support for auto reconnection in favor above the fail one reconnection that was previously implemented as well as setting connection timezone.