Extended support for HTTP response codes
Added nearly every standard (No WebDAV) HTTP response code to the server. I don't see the use case for most of them but if JDC is ever used in the niche cases where some of them are necessary they can be used to return a resource by assigning the error file in the error.json configuration file.